refactor(merk/proofs): hoist shared aggregate-on-range helpers to aggregate_common
`SubtreeClassification`, `classify_subtree`, `key_strictly_inside`, and
the `NULL_HASH` const were byte-identical between
`merk/src/proofs/query/aggregate_count.rs` and
`merk/src/proofs/query/aggregate_sum.rs` — the sum-side copy's
doc-comment even read "Mirrors the count side exactly" and "Identical
logic to the count side".

This is a security-relevant duplication: the range-bound classification
is what the verifier uses to decide whether a subtree is Disjoint,
Contained, or Boundary. If the count and sum copies ever drift (say,
one accepts a subtree the other rejects), the two aggregate paths
would disagree on what a "valid" proof shape looks like — an
attacker who finds a proof that one accepts and the other rejects
has a malleability surface to exploit.

Extract them into a new `merk::proofs::query::aggregate_common`
module so the contract has exactly one definition. Both
aggregate-count and aggregate-sum modules now import the four items
from `aggregate_common` (`pub(super)` so external callers can't
reach them).

The bound math is provably independent of the aggregate flavor —
it depends only on the key window relative to the inner range —
so this consolidation can't change verifier behavior. The most
comprehensive existing inline doc-comments (the count-side
"see the inline proofs" rationale and the per-arm bound-math
explanations) are preserved verbatim in the new home.

Tests: 3199 / 0 fail (unchanged). Clippy clean.
